pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gearmeshing-ai"
version = "0.0.0"
description = "🤖 Enterprise AI bases development team."
authors = [{ name = "Chisanan232", email = "chi10211201@cycu.org.tw" }]
requires-python = ">=3.13,<4.0"
readme = "README.md"
license = { file = "LICENSE" }
keywords = [
    "ai",
    "ai agents",
    "ai agent team",
    "ai-assistents",
    "private",
    "enterprise",
    "controllable",
    "efficiant",
    "monitoring",
    "tracing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    # Web framework
    "fastapi>=0.128.0",
    "sse-starlette>=3.2.0",
    "uvicorn[standard]>=0.40.0",
    # Data validation for web framework
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    # Command line interface
    "typer>=0.21.1",
    # MCP SDK
    "mcp>=1.26.0",
    # MCP Contextforge Gateway for MCP servers
    "mcp-contextforge-gateway>=0.9.0",
    # HTTP client
    "httpx>=0.28.1",
    # AI agent framework
    "pydantic-ai>=1.48.0",
    # AI agent workflow framework
    "langgraph>=1.0.7",
    # Scheduler framework as a workflow engine
    "temporalio>=1.20.0",
    # Extranal service operations
    "clickup-mcp-server>=0.1.0",
    "slack-mcp>=0.1.0",
]

[project.scripts]
gearmeshing-ai = "gearmeshing_ai.command_line.app:main_entry"

[project.urls]
Homepage = "https://github.com/Chisanan232/GearMeshing-AI"
#Documentation = "https://chisanan232.github.io/GearMeshing-AI/intro"
Repository = "https://github.com/Chisanan232/GearMeshing-AI"
#Changelog = "<the release note page in documentation or GitHub release info>"
"Project Management" = "https://app.clickup.com/9018752317/v/f/90185307863/90182761188"

[dependency-groups]
# Dependencies for the development
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=5.0.0",
    "pytest-rerunfailures>=16.0.1",
    "coverage>=7.10.6",
    "testcontainers>=4.14.0",
]
# Dependencies for the pre-commit CI in development environment
pre-commit-ci = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "ruff>=0.14.14",
]
# Optional monitoring and observability
monitoring = [
    "logfire[fastapi,sqlalchemy]>=4.21.0",
]

[tool.uv]
default-groups = [
    "dev",
    "pre-commit-ci",
]

[tool.hatch.build.targets.sdist]
include = ["gearmeshing_ai/"]
packages = ["gearmeshing_ai"]
artifacts = ["gearmeshing_ai/py.typed"]

[tool.hatch.build.targets.wheel]
include = ["gearmeshing_ai/"]
packages = ["gearmeshing_ai"]
artifacts = ["gearmeshing_ai/py.typed"]
